requirements.txt
openai>=1.30.0
pandas>=2.2.0
openpyxl>=3.1.2
requests>=2.31.0
schedule>=1.2.1
python-dotenv>=1.0.1
duckduckgo-search>=6.1.0
tavily-python>=0.5.0
Flask>=3.0.0
gunicorn>=22.0.0
